Creating products work

+++ src/models/Product.php
@@ -18,7 +18,7 @@ class Product
 
     private $required = ['catalog_cwid', 'customer_desc', 'quantity', 'price', 'cost', 'opp_cwid', 'locationID', 'businessUnitId'];
 
-    public function __construct(int $catalog_cwid, string $customer_desc, int $quantity, int $price, int $cost, int $opp_cwid, int $locationID, int $businessUnitId)
+    public function __construct(int $catalog_cwid, string $customer_desc, int $quantity, float $price, float $cost, int $opp_cwid, int $locationID, int $businessUnitId)
     {
         $this->catalog_cwid = $catalog_cwid;
         $this->customer_desc = $customer_desc;
